+++ src/proj_i.c
@@ -325,7 +325,8 @@ int rowcol_to_latlon_i( float row, float col, float *lat, float *lon,
          break;
 
       default:
-         printf("Error in rowcol_to_latlon_i: bad projection: %d\n", proj );
+         printf("Error in rowcol_to_latlon_i: bad projection: %d\n",
+		(int)proj );
          break;
    }
 
@@ -596,7 +597,8 @@ int latlon_to_rowcol_i( float lat, float lon, float *row, float *col,
          break;
 
       default:
-         printf("Error in latlon_to_rowcol_i: bad projection: %d\n", proj );
+         printf("Error in latlon_to_rowcol_i: bad projection: %d\n",
+		(int)proj );
          break;
    }
 
@@ -662,7 +664,7 @@ float proj_resolution( struct projection *proj )
          break;
 
       default:
-         printf("Error in proj_resolution: bad projection: %d\n", proj );
+         printf("Error in proj_resolution: bad projection: %d\n", (int)proj );
          break;
    }
